finished porting lib compat for now
