+++ felt-item-api/src/main/java/io/github/feltmc/feltapi/mixin/item/ItemMixin.java
@@ -1,6 +1,5 @@
 package io.github.feltmc.feltapi.mixin.item;
 
-import io.github.fabricators_of_create.porting_lib.item.CustomMaxCountItem;
 import io.github.feltmc.feltapi.api.item.extensions.IsDamageableItem;
 import io.github.feltmc.feltapi.api.item.extensions.MiscExtension;
 import net.minecraft.item.Item;
@@ -14,28 +13,10 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(Item.class)
 public abstract class ItemMixin {
-    @Shadow public abstract boolean isDamageable();
-
-    @Shadow public abstract int getMaxCount();
-
     @Inject(method = "isIn", at = @At("HEAD"), cancellable = true)
     private void injectGetGroups(ItemGroup group, CallbackInfoReturnable<Boolean> cir){
         if(this instanceof MiscExtension extension){
             if (extension.getGroups().stream().anyMatch(g -> g == group)) cir.setReturnValue(true);
-        }
-
-    }
-
-    @Inject(method = "isEnchantable", at = @At("HEAD"), cancellable = true)
-    private void injectIsEnchantable(ItemStack stack, CallbackInfoReturnable<Boolean> cir){
-        if (this instanceof CustomMaxCountItem count){
-            if (this instanceof IsDamageableItem extension){
-                cir.setReturnValue(count.getItemStackLimit(stack) == 1 && extension.isDamageable(stack));
-                return;
-            }
-            cir.setReturnValue(count.getItemStackLimit(stack) == 1 && this.isDamageable());
-        } else if (this instanceof IsDamageableItem extension){
-            cir.setReturnValue(this.getMaxCount() == 1 && extension.isDamageable(stack));
         }
 
     }
